+++ harvey/globals.py
@@ -1,4 +1,3 @@
-import ipaddress
 import os
 
 
@@ -20,13 +19,6 @@ class Global:
         'test',
         'deploy',
         'full',
-    ]
-    # TODO: Get these directly from the API instead of a static list
-    GITHUB_WEBHOOK_IP_RANGES = [
-        '192.30.252.0/22',
-        '185.199.108.0/22',
-        '140.82.112.0/20',
-        '143.55.64.0/20',
     ]
 
     @staticmethod
@@ -58,23 +50,3 @@ class Global:
     def repo_commit_id(webhook):
         """Return the repo's id from the webhook JSON."""
         return webhook['commits'][0]['id']
-
-    @staticmethod
-    def github_webhook_ip_ranges():
-        """Returns a list of public IP addresses GitHub sends webhooks from.
-
-        TODO: Refactor this so it's not a static list but a list pulled from GitHub
-        and stored in redis or a database:
-
-        https://docs.github.com/en/github/authenticating-to-github/keeping-your-account-and-data-secure/about-githubs-ip-addresses
-        """
-        ip_address_list = [
-            str(ip_address)
-            for ip_range in Global.GITHUB_WEBHOOK_IP_RANGES
-            for ip_address in ipaddress.IPv4Network(ip_range)
-        ]
-
-        # Keep localhost included for testing
-        ip_address_list.append('127.0.0.1')
-
-        return ip_address_list