+++ test/unit/test_webhooks.py
@@ -80,3 +80,13 @@ def test_webhook_originated_outside_github(mock_webhook_object):
         'success': False,
     }
     assert webhook[1] == 422
+
+
+def test_get_github_ip_address():
+    """Assert that localhost is in the results and that the length
+    exceeds some arbitrarily large number as GitHub has MANY ip addresses.
+    """
+    result = Webhook.get_github_ip_address()
+
+    assert '127.0.0.1' in result
+    assert len(result) > 1000  # Actual number is close to 1900